+++ enterprise/frontend/src/metabase-enterprise/api/index.ts
@@ -1,6 +1,6 @@
 export * from "./audit-info";
 export * from "./billing-info";
-export * from "./gsheets";
+export * from "./gdrive";
 export * from "./query-validation";
 export * from "./scim";
 export * from "./upload-management";
