+++ enterprise/frontend/src/metabase-enterprise/google_drive/GdriveConnectionModal.tsx
@@ -1,10 +1,7 @@
 import { type FormEvent, useState } from "react";
-import { useLocation } from "react-use";
-import { match } from "ts-pattern";
 import { jt, t } from "ttag";
 
 import { reloadSettings } from "metabase/admin/settings/settings";
-import { skipToken, useGetDatabaseQuery } from "metabase/api";
 import { useSetting } from "metabase/common/hooks";
 import { CopyButton } from "metabase/components/CopyButton";
 import { useDispatch, useSelector } from "metabase/lib/redux";
@@ -14,8 +11,6 @@ import {
   Button,
   Center,
   Flex,
-  Icon,
-  Menu,
   Modal,
   Stack,
   Text,
@@ -27,147 +22,10 @@ import {
   useSaveGsheetsFolderLinkMutation,
 } from "metabase-enterprise/api";
 
-import { trackSheetConnectionClick, trackSheetImportClick } from "./analytics";
+import { trackSheetImportClick } from "./analytics";
 import disconnectIllustration from "./disconnect.svg?component";
 
-export function GsheetConnectButton() {
-  const url = useLocation();
-  const databaseId = /databases\/(\d+)/.exec(url.pathname ?? "")?.[1];
-
-  const { data: databaseInfo } = useGetDatabaseQuery(
-    databaseId ? { id: Number(databaseId) } : skipToken,
-  );
-
-  const isDwh = databaseInfo?.is_attached_dwh;
-
-  const [showModal, setShowModal] = useState(false);
-
-  const gSheetsSetting = useSetting("gsheets");
-  const gSheetsEnabled = useSetting("show-google-sheets-integration");
-
-  const userIsAdmin = useSelector(getUserIsAdmin);
-
-  const shouldGetServiceAccount = isDwh && gSheetsEnabled && userIsAdmin;
-
-  const { data: { email: serviceAccountEmail } = {} } =
-    useGetServiceAccountQuery(shouldGetServiceAccount ? undefined : skipToken);
-
-  if (
-    !gSheetsEnabled ||
-    !gSheetsSetting ||
-    !userIsAdmin ||
-    !serviceAccountEmail ||
-    !isDwh
-  ) {
-    return null;
-  }
-
-  const { status } = gSheetsSetting;
-
-  const buttonText = match(status)
-    .with("not-connected", () => t`Connect Google Sheets`)
-    .with("loading", () => t`Google Sheets connecting...`)
-    .with("complete", () => t`Disconnect`)
-    .otherwise(() => t`Google Sheets`);
-
-  return (
-    <Flex align="center" gap="sm">
-      {status === "complete" && (
-        <Flex align="center" gap="xs">
-          <Icon name="google_sheet" />
-          <Text>{t`Connected to Google Sheets`}</Text>
-        </Flex>
-      )}
-      <Text>{" · "}</Text>
-      <Button
-        p={0}
-        variant="subtle"
-        onClick={() => {
-          setShowModal(true);
-          trackSheetConnectionClick({ from: "db-page" });
-        }}
-        disabled={status === "loading"}
-        leftSection={
-          status === "complete" ? undefined : <Icon name="google_sheet" />
-        }
-      >
-        {buttonText}
-      </Button>
-      <GsheetConnectionModal
-        isModalOpen={showModal}
-        onClose={() => setShowModal(false)}
-        reconnect={false}
-      />
-    </Flex>
-  );
-}
-
-export function GsheetMenuItem({ onClick }: { onClick: () => void }) {
-  const gSheetsSetting = useSetting("gsheets");
-  const gSheetsEnabled = useSetting("show-google-sheets-integration");
-
-  const userIsAdmin = useSelector(getUserIsAdmin);
-  const { data: { email: serviceAccountEmail } = {} } =
-    useGetServiceAccountQuery();
-
-  if (
-    !gSheetsEnabled ||
-    !gSheetsSetting ||
-    !userIsAdmin ||
-    !serviceAccountEmail
-  ) {
-    return null;
-  }
-
-  const { status } = gSheetsSetting;
-
-  const handleClick = () => {
-    trackSheetConnectionClick({ from: "left-nav" });
-    onClick();
-  };
-
-  const buttonText = match(status)
-    .with("not-connected", () => t`Connect Google Sheets`)
-    .with("loading", () => t`Google Sheets`)
-    .with("complete", () => t`Google Sheets`)
-    .otherwise(() => t`Google Sheets`);
-
-  const helperText = match(status)
-    .with("not-connected", () => null)
-    .with("loading", () => t`Importing files...`)
-    .with("complete", () => t`Connected`)
-    .otherwise(() => null);
-
-  return (
-    <Menu.Item onClick={handleClick} disabled={status === "loading"}>
-      <Flex gap="sm" align="flex-start" justify="space-between" w="100%">
-        <Flex>
-          <Icon name="google_sheet" mt="xs" mr="sm" />
-          <div>
-            <Text
-              fw="bold"
-              c={status === "loading" ? "text-light" : "text-dark"}
-            >
-              {buttonText}
-            </Text>
-            {helperText && (
-              <Text size="sm" c="text-medium">
-                {helperText}
-              </Text>
-            )}
-          </div>
-        </Flex>
-        {status === "complete" && (
-          <Button variant="subtle" onClick={handleClick}>
-            {t`Add New`}
-          </Button>
-        )}
-      </Flex>
-    </Menu.Item>
-  );
-}
-
-export function GsheetConnectionModal({
+export function GdriveConnectionModal({
   isModalOpen,
   onClose,
   reconnect,
