+++ enterprise/frontend/src/metabase-enterprise/google_drive/index.ts
@@ -1,0 +1,4 @@
+export * from "./GdriveConnectionModal";
+export * from "./GdriveSidebarMenuItem";
+export * from "./GdriveSyncStatus";
+export * from "./GdriveDbMenu";
