--- enterprise/frontend/src/metabase-enterprise/google_sheets/index.ts
@@ -1,2 +1,0 @@
-export * from "./GoogleSheetManagement";
-export * from "./GsheetsSyncStatus";
