+++ enterprise/frontend/src/metabase-enterprise/upload_management/index.ts
@@ -2,11 +2,11 @@ import { PLUGIN_UPLOAD_MANAGEMENT } from "metabase/plugins";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 
 import {
-  GsheetConnectButton,
-  GsheetConnectionModal,
-  GsheetMenuItem,
-  GsheetsSyncStatus,
-} from "../google_sheets";
+  GdriveConnectionModal,
+  GdriveDbMenu,
+  GdriveSidebarMenuItem,
+  GdriveSyncStatus,
+} from "../google_drive";
 
 import { UploadManagementTable } from "./UploadManagementTable";
 
@@ -15,8 +15,8 @@ if (hasPremiumFeature("upload_management")) {
 }
 
 if (hasPremiumFeature("hosting") && hasPremiumFeature("attached_dwh")) {
-  PLUGIN_UPLOAD_MANAGEMENT.GsheetConnectionModal = GsheetConnectionModal;
-  PLUGIN_UPLOAD_MANAGEMENT.GsheetMenuItem = GsheetMenuItem;
-  PLUGIN_UPLOAD_MANAGEMENT.GsheetsSyncStatus = GsheetsSyncStatus;
-  PLUGIN_UPLOAD_MANAGEMENT.GsheetConnectButton = GsheetConnectButton;
+  PLUGIN_UPLOAD_MANAGEMENT.GdriveConnectionModal = GdriveConnectionModal;
+  PLUGIN_UPLOAD_MANAGEMENT.GdriveSidebarMenuItem = GdriveSidebarMenuItem;
+  PLUGIN_UPLOAD_MANAGEMENT.GdriveSyncStatus = GdriveSyncStatus;
+  PLUGIN_UPLOAD_MANAGEMENT.GdriveDbMenu = GdriveDbMenu;
 }
