+++ frontend/src/metabase/nav/containers/MainNavbar/tests/enterprise.unit.spec.tsx
@@ -1,11 +1,11 @@
-import { setupGsheetsServiceAccountEndpoint } from "__support__/server-mocks";
+import { setupGdriveServiceAccountEndpoint } from "__support__/server-mocks";
 import { screen } from "__support__/ui";
 import { createMockUser } from "metabase-types/api/mocks";
 
 import { type SetupOpts, setup } from "./setup";
 
 const setupEnterprise = (opts?: SetupOpts) => {
-  setupGsheetsServiceAccountEndpoint();
+  setupGdriveServiceAccountEndpoint();
   return setup({ ...opts, hasEnterprisePlugins: true });
 };
