+++ frontend/test/__support__/server-mocks/index.ts
@@ -15,7 +15,7 @@ export * from "./dataset";
 export * from "./email";
 export * from "./field";
 export * from "./group";
-export * from "./gsheets";
+export * from "./gdrive";
 export * from "./impersonation";
 export * from "./model-indexes";
 export * from "./native-query-snippet";
